requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torch-geometric>=2.0.0
numpy>=1.19.2
pandas>=1.2.4
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
wandb>=0.12.0
pyyaml>=5.4.1
tqdm>=4.62.3
pytorch-lightning>=1.5.0
albumentations>=1.1.0

